feat(tests) : improved tests by adding jest as testing framework
